requirements.txt
# Core ML & DL
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Vision & Transfer Learning
timm>=0.9.7

# Experiments & Tracking
mlflow>=2.5.0

# Preprocessing & Augmentation
Pillow>=9.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web App
streamlit>=1.26.0

# Data & Datasets
medmnist>=3.0.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.65.0

# Development
jupyter>=1.0.0
ipython>=8.14.0
